+++ src/types/theme.ts
@@ -1,6 +1,8 @@
 import type { OverrideProperties } from 'type-fest';
 import type { TextMateThemingRule, TokenColorizationSetting } from './vscode.ts';
 
+export type Hex = '0' | '1' | '2' | '3' | '4' | '5' | '6' | '7' | '8' | '9' | 'a' | 'b' | 'c' | 'd' | 'e' | 'f';
+
 export type Color = `#${string}`;
 
 export type ThemeConfig = {