+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 0.4.0
+### feature
+- ⚛️ add new variant theme based on Atom One Dark
+- 🔢 modify editor line's highlight background
+
 ## 0.3.0
 ### feature
 - 📚 markdown preview custom blockquote style